Mapping corrected

+++ src/main/java/com/spring/gradle/defect/repository/DefectRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.spring.gradle.defect.entity.Defect;
 
 @Repository
-public interface DefectRepository extends JpaRepository<Defect, Long> {
+public interface DefectRepository extends JpaRepository<Defect,Integer> {
 
 }
